refactor(jmap): rename methods/ to _methods/ to signal internal API
